requirements.txt
supabase>=2.0.0
pandas>=2.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
